accounts-microservice/requirements.txt
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
dj-database-url==1.0.0
django-cors-headers==3.11.0
Django==4.0.3
django-filter==22.1
djwto @ git+https://github.com/hr-curtissimo/djwto.git
factory-boy==3.2.1
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
markdown==3.4.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
redis==4.3.4
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
whitenoise==5.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
pika==1.3.1
